Rename Eclipse V-SUM project, namespaces and IDs
