+++ bundles/tools.vitruv.adaptors.eclipse.vsum/.classpath
