+++ bundles/tools.vitruv.adaptors.eclipse.vsum/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>tools.vitruv.framework.vsum.ui</name>
+	<name>tools.vitruv.adapters.eclipse.vsum</name>
 	<comment></comment>
 	<projects>
 	</projects>
