+++ bundles/tools.vitruv.adaptors.eclipse.vsum/.settings/org.eclipse.core.resources.prefs
