+++ bundles/tools.vitruv.adaptors.eclipse.vsum/META-INF/MANIFEST.MF
@@ -1,10 +1,10 @@
 Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
-Bundle-Name: Vitruv Framework UI Virtual Single Underlying Model
-Bundle-SymbolicName: tools.vitruv.framework.vsum.ui;singleton:=true
-Automatic-Module-Name: tools.vitruv.framework.vsum.ui
+Bundle-Name: Vitruv Eclipse Adapter for V-SUM Creation
+Bundle-SymbolicName: tools.vitruv.adapters.eclipse.vsum;singleton:=true
+Automatic-Module-Name: tools.vitruv.adapters.eclipse.vsum
 Bundle-Version: 2.1.0.qualifier
-Bundle-Activator: tools.vitruv.framework.vsum.ui.Activator
+Bundle-Activator: tools.vitruv.adapters.eclipse.vsum.Activator
 Require-Bundle: org.eclipse.ui,
  org.eclipse.core.runtime,
  org.eclipse.core.resources,
@@ -17,5 +17,5 @@ Require-Bundle: org.eclipse.ui,
 Bundle-RequiredExecutionEnvironment: JavaSE-11
 Bundle-ActivationPolicy: lazy
 Bundle-Vendor: vitruv.tools
-Export-Package: tools.vitruv.framework.vsum.ui,
- tools.vitruv.framework.vsum.ui.wizard
+Export-Package: tools.vitruv.adapters.eclipse.vsum,
+ tools.vitruv.adapters.eclipse.vsum.wizard
