+++ bundles/tools.vitruv.adaptors.eclipse.vsum/build.properties
