+++ bundles/tools.vitruv.adaptors.eclipse.vsum/images/vitruvian.png
