+++ bundles/tools.vitruv.adaptors.eclipse.vsum/plugin.xml
@@ -4,14 +4,14 @@
    <extension
          point="org.eclipse.ui.newWizards">
       <category
-            id="tools.vitruv.framework.vsum.ui.Vitruv"
+            id="tools.vitruv.adapters.eclipse.vsum"
             name="Vitruv">
       </category>
       <wizard
-            category="tools.vitruv.framework.vsum.ui.Vitruv"
-            class="tools.vitruv.framework.vsum.ui.wizard.CreateVsumWizard"
+            category="tools.vitruv.adapters.eclipse.vsum"
+            class="tools.vitruv.adapters.eclipse.vsum.wizard.CreateVsumWizard"
             icon="images/vitruvian.png"
-            id="tools.vitruv.framework.vsum.ui.newvitruvwizard"
+            id="tools.vitruv.adapters.eclipse.vsum.newvitruvwizard"
             name="Vitruv Project">
       </wizard>
    </extension>
@@ -21,9 +21,9 @@
             allPopups="false"
             locationURI="toolbar:org.eclipse.ui.main.toolbar">
          <toolbar
-               id="tools.vitruv.framework.vsum.ui.toolbar">
+               id="tools.vitruv.adapters.eclipse.vsum.toolbar">
             <command
-                  commandId="tools.vitruv.framework.vsum.ui.vitruvwizardcommand"
+                  commandId="tools.vitruv.adapters.eclipse.vsum.vitruvwizardcommand"
                   icon="images/vitruvian.png"
                   style="push">
             </command>
@@ -31,11 +31,11 @@
       </menuContribution>
    </extension>
    <extension
-         id="tools.vitruv.framework.vsum.ui.vitruvwizardcommand"
+         id="tools.vitruv.adapters.eclipse.vsum.vitruvwizardcommand"
          point="org.eclipse.ui.commands">
       <command
-            defaultHandler="tools.vitruv.framework.vsum.ui.CreateVsumButtonHandler"
-            id="tools.vitruv.framework.vsum.ui.vitruvwizardcommand"
+            defaultHandler="tools.vitruv.adapters.eclipse.vsum.CreateVsumButtonHandler"
+            id="tools.vitruv.adapters.eclipse.vsum.vitruvwizardcommand"
             name="VitruvWizard">
       </command>
    </extension>
