+++ bundles/tools.vitruv.adaptors.eclipse.vsum/src/tools/vitruv/adapters/eclipse/vsum/CreateVsumButtonHandler.xtend
@@ -1,11 +1,11 @@
-package tools.vitruv.framework.vsum.ui
+package tools.vitruv.adapters.eclipse.vsum
 
 import org.eclipse.core.commands.AbstractHandler
 import org.eclipse.core.commands.ExecutionEvent
 import org.eclipse.core.commands.ExecutionException
 import org.eclipse.ui.PlatformUI
 import org.eclipse.jface.wizard.WizardDialog
-import tools.vitruv.framework.vsum.ui.wizard.CreateVsumWizard
+import tools.vitruv.adapters.eclipse.vsum.wizard.CreateVsumWizard
 
 class CreateVsumButtonHandler extends AbstractHandler {
